requirements.txt
asgiref==3.11.0
astroid==3.3.11
colorama==0.4.6
dill==0.4.0
Django==6.0
django-ckeditor==6.7.3
django-js-asset==3.1.2
isort==6.1.0
mccabe==0.7.0
pillow==12.0.0
platformdirs==4.5.1
pylint==3.3.9
pylint-django==2.6.1
pylint-plugin-utils==0.9.0
sqlparse==0.5.5
tomlkit==0.13.3
tzdata==2025.3
